Extract `tool_utils` package
