+++ app/src/jvmMain/kotlin/diy/lingerie/dump_tool/main.kt
@@ -9,18 +9,11 @@ import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.enum
 import com.github.ajalt.clikt.parameters.types.path
 import diy.lingerie.geometry.LineSegment
-import diy.lingerie.geometry.Point
 import diy.lingerie.geometry.Ray
 import diy.lingerie.geometry.curves.PrimitiveCurve
-import diy.lingerie.geometry.splines.OpenSpline
 import diy.lingerie.geometry.splines.Spline
-import diy.lingerie.geometry.svg.importSvgPath
-import diy.lingerie.geometry.transformations.Transformation
-import diy.lingerie.simple_dom.svg.SvgLine
-import diy.lingerie.simple_dom.svg.SvgPath
-import diy.lingerie.simple_dom.svg.SvgRectangle
 import diy.lingerie.simple_dom.svg.SvgRoot
-import diy.lingerie.simple_dom.svg.SvgShape
+import diy.lingerie.tool_utils.RecognizedShape
 import java.nio.file.Path
 import kotlin.io.path.reader
